fix: solve errors in components
