+++ app/admin/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { Box, Text } from '@mantine/core';
 import React from 'react';
 const location = {
