+++ app/blogs/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import {
   Button,
   Container,
