+++ app/login/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React from 'react';
 import Auth from '../../src/components/Auth/Login';
