+++ app/register/page.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 import Register from '../../src/components/Auth/Register';
